Moved Filter to components folder
